added overall score filter for diseases table
