+++ client/client/app/components/ngbTargetPanel/ngbDiseasesTab/ngbDiseasesTargetsPanel/ngbDiseasesTargetsTableFilter/ngbTargetsFilterInput/ngbTargetsFilterInput.controller.js
@@ -2,10 +2,15 @@ export default class ngbTargetsFilterInputController {
     prevValue;
     value;
 
-    constructor(dispatcher, ngbDiseasesTargetsPanelService) {
+    constructor($scope, dispatcher, ngbDiseasesTargetsPanelService) {
         this.dispatcher = dispatcher;
         this.ngbDiseasesTargetsPanelService = ngbDiseasesTargetsPanelService;
         this.prevValue = this.value = this.ngbDiseasesTargetsPanelService.filterInfo[this.column.field];
+
+        this.dispatcher.on('target:diseases:targets:filters:reset', this.resetFilters.bind(this));
+        $scope.$on('$destroy', () => {
+            dispatcher.removeListener('target:diseases:targets:filters:reset', this.resetFilters.bind(this));
+        });
     }
 
     static get UID() {
@@ -18,5 +23,10 @@ export default class ngbTargetsFilterInputController {
             this.ngbDiseasesTargetsPanelService.setFilter(this.column.field, this.value);
             this.dispatcher.emit('target:diseases:targets:filters:changed');
         }
+    }
+
+    resetFilters() {
+        this.value = undefined;
+        this.prevValue = undefined;
     }
 }
