+++ client/client/app/components/ngbTargetPanel/ngbIdentificationsTab/ngbDiseasesPanel/ngbDiseasesTable/ngbDiseasesTableFilter/ngbDiseasesFilterInput/index.js
@@ -1,0 +1,9 @@
+import angular from 'angular';
+
+import component from './ngbDiseasesFilterInput.component';
+import controller from './ngbDiseasesFilterInput.controller';
+
+export default angular.module('ngbDiseasesFilterInput', [])
+    .controller(controller.UID, controller)
+    .component('ngbDiseasesFilterInput', component)
+    .name;
