+++ client/client/app/components/ngbTargetPanel/ngbIdentificationsTab/ngbDiseasesPanel/ngbDiseasesTable/ngbDiseasesTableFilter/ngbDiseasesFilterInput/ngbDiseasesFilterInput.component.js
@@ -1,0 +1,9 @@
+import controller from './ngbDiseasesFilterInput.controller';
+
+export default  {
+    bindings: {
+        column: '<'
+    },
+    controller: controller.UID,
+    template: require('./ngbDiseasesFilterInput.tpl.html')
+};
