+++ client/client/app/components/ngbTargetPanel/ngbIdentificationsTab/ngbDiseasesPanel/ngbDiseasesTable/ngbDiseasesTableFilter/ngbDiseasesFilterInput/ngbDiseasesFilterInput.tpl.html
@@ -1,0 +1,6 @@
+<input class="ui-grid-filter-input ngb-filter-input"
+       ng-blur="$ctrl.apply()"
+       ng-key-bind="{enter: '$ctrl.apply()'}"
+       ng-model="$ctrl.value"
+       style="margin-bottom: 1px"
+       type="text">
