+++ client/client/app/components/ngbTargetPanel/ngbIdentificationsTab/ngbDiseasesPanel/ngbDiseasesTable/ngbDiseasesTableFilter/ngbDiseasesFilterList/ngbDiseasesFilterList.component.js
@@ -1,0 +1,9 @@
+import controller from './ngbDiseasesFilterList.controller';
+
+export default  {
+    bindings: {
+        column: '<'
+    },
+    controller: controller.UID,
+    template: require('./ngbDiseasesFilterList.tpl.html')
+};
